requirements.awsprod.txt
absl-py==1.2.0
astroid==2.11.5
astunparse==1.6.3
aws-cfn-bootstrap==2.0
bidict==0.22.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
Cython==0.29.30
dataclasses==0.6
dill==0.3.5.1
dm-tree==0.1.7
docutils==0.14
etils==0.6.0
Flask==2.1.2
Flask-SocketIO==5.2.0
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gevent==21.12.0
gevent-websocket==0.10.1
gin-config==0.5.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.14.1
google-cloud-core==2.3.2
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
h5py==3.7.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kaggle==1.5.12
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
libclang==14.0.1
lockfile==0.11.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
networkx==2.6.3
nltk==3.7
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
platformdirs==2.5.2
promise==2.3
proto-plus==1.20.6
protobuf==3.19.4
psutil==5.9.1
py-cpuinfo==8.0.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pylint==2.14.1
pyparsing==3.0.9
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
python-engineio==4.3.2
python-slugify==6.1.2
python-socketio==5.6.0
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
seqeval==1.2.2
simplejson==3.2.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-addons==0.15.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
tensorflow-model-optimization==0.7.3
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
typed-ast==1.5.4
typeguard==2.13.3
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.10
uWSGI==2.0.20
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
